Fixing Jest tests
