+++ package.json
@@ -32,8 +32,6 @@
     "react-router-dom": "^5.0.1",
     "react-scripts": "3.1.1",
     "redux": "^4.0.4",
-    "react-test-renderer": "^16.9.0",
-    "redux": "^4.0.4",
     "yup": "^0.27.0"
   },
   "scripts": {
