+++ src/Components/Avatar/utils.js
@@ -6,13 +6,17 @@ function getBase64(img, callback) {
     reader.readAsDataURL(img);
 }
 
+function isLowerThan2m(size) {
+    return size / 1024 / 1024 < 2;
+}
+
 function beforeUpload(file) {
     const { type, size } = file;
     const isJpgOrPng = type === 'image/jpeg' || type === 'image/png';
     if (!isJpgOrPng) {
         message.error('You can only upload JPG/PNG file!');
     }
-    const isLt2M = size / 1024 / 1024 < 2;
+    const isLt2M = isLowerThan2m(size);
     if (!isLt2M) {
         message.error('Image must smaller than 2MB!');
     }
